fix: moved parseQuery to common-utils, added check for saved search filter schema to ensure it is renderable on UI. Moved check into utils
