+++ packages/api/openapi.json
@@ -3071,74 +3071,28 @@
         }
       },
       "SavedSearchFilter": {
-        "description": "A single structured pinned filter applied to the search. Either a text-condition filter (Lucene or SQL) or a structured SQL AST comparison.",
-        "oneOf": [
-          {
-            "type": "object",
-            "required": [
-              "type",
-              "condition"
+        "type": "object",
+        "required": [
+          "condition"
+        ],
+        "description": "A single pinned filter applied to the search. Filters must use the SQL predicate form the UI produces so they render as a sidebar facet: `<column> IN ('<value1>', ...)` (or `NOT IN` / `BETWEEN`). Conditions that are not in this form are rejected, since they would not appear in the sidebar.",
+        "properties": {
+          "type": {
+            "type": "string",
+            "enum": [
+              "sql"
             ],
-            "properties": {
-              "type": {
-                "type": "string",
-                "enum": [
-                  "lucene",
-                  "sql"
-                ],
-                "description": "Language of the condition expression."
-              },
-              "condition": {
-                "type": "string",
-                "maxLength": 8192,
-                "description": "Filter expression in the language given by type.",
-                "example": "SeverityText:ERROR"
-              }
-            }
+            "default": "sql",
+            "description": "Always `sql`. Only SQL predicate filters render in the sidebar.",
+            "example": "sql"
           },
-          {
-            "type": "object",
-            "required": [
-              "type",
-              "operator",
-              "left",
-              "right"
-            ],
-            "properties": {
-              "type": {
-                "type": "string",
-                "enum": [
-                  "sql_ast"
-                ],
-                "description": "Marks a structured SQL AST comparison."
-              },
-              "operator": {
-                "type": "string",
-                "enum": [
-                  "=",
-                  "<",
-                  ">",
-                  "!=",
-                  "<=",
-                  ">="
-                ],
-                "description": "Comparison operator."
-              },
-              "left": {
-                "type": "string",
-                "maxLength": 8192,
-                "description": "Left-hand column or expression.",
-                "example": "ServiceName"
-              },
-              "right": {
-                "type": "string",
-                "maxLength": 8192,
-                "description": "Right-hand literal or expression.",
-                "example": "'checkout'"
-              }
-            }
+          "condition": {
+            "type": "string",
+            "maxLength": 8192,
+            "description": "SQL predicate applied to the search, in `<column> IN (...)` form.",
+            "example": "ServiceName IN ('checkout', 'payments')"
           }
-        ]
+        }
       },
       "SavedSearch": {
         "type": "object",
@@ -3210,8 +3164,8 @@
             },
             "example": [
               {
-                "type": "lucene",
-                "condition": "SeverityText:ERROR"
+                "type": "sql",
+                "condition": "ServiceName IN ('checkout', 'payments')"
               }
             ]
           },
@@ -3305,8 +3259,8 @@
             },
             "example": [
               {
-                "type": "lucene",
-                "condition": "SeverityText:ERROR"
+                "type": "sql",
+                "condition": "ServiceName IN ('checkout', 'payments')"
               }
             ]
           }
@@ -4567,6 +4521,7 @@
         "properties": {
           "name": {
             "type": "string",
+            "maxLength": 1024,
             "description": "Webhook name. Must be unique per service within the team.",
             "example": "Production Alerts"
           },
@@ -4583,24 +4538,29 @@
           "url": {
             "type": "string",
             "format": "uri",
+            "maxLength": 2048,
             "description": "Webhook destination URL.",
             "example": "https://hooks.slack.com/services/T00000000/B00000000/XXXXXXXXXXXXXXXXXXXXXXXX"
           },
           "description": {
             "type": "string",
+            "maxLength": 2048,
             "description": "Webhook description, shown in the UI.",
             "example": "Sends critical alerts to the #incidents channel"
           },
           "body": {
             "type": "string",
+            "maxLength": 16384,
             "description": "Optional request body template. Only for generic/incidentio; rejected for slack.",
             "example": "{\"alert\": \"{{title}}\", \"severity\": \"{{level}}\"}"
           },
           "headers": {
             "type": "object",
-            "description": "Write-only. Custom HTTP headers sent with the webhook request. Never returned on read. Only for generic/incidentio; rejected for slack.",
+            "maxProperties": 100,
+            "description": "Write-only. Custom HTTP headers sent with the webhook request. Never returned on read. Only for generic/incidentio; rejected for slack. Each value is capped at 4096 characters.",
             "additionalProperties": {
-              "type": "string"
+              "type": "string",
+              "maxLength": 4096
             },
             "example": {
               "Authorization": "Bearer secret-token"
@@ -4608,9 +4568,11 @@
           },
           "queryParams": {
             "type": "object",
-            "description": "Write-only. Query parameters appended to the webhook URL. Never returned on read. Only for generic/incidentio; rejected for slack.",
+            "maxProperties": 100,
+            "description": "Write-only. Query parameters appended to the webhook URL. Never returned on read. Only for generic/incidentio; rejected for slack. Each value is capped at 4096 characters.",
             "additionalProperties": {
-              "type": "string"
+              "type": "string",
+              "maxLength": 4096
             }
           }
         }
